Phase 3: привести pstring/pvector/pmap/pjson к uintptr_t

+++ pmap.h
@@ -13,6 +13,8 @@
 //   - K должен поддерживать operator== (равенство).
 //   - pmap_data<K, V> тривиально копируем.
 //
+// Phase 3: size и capacity обновлены до uintptr_t через pvector_data (задача 3.3).
+//
 // Использование:
 //   pmap_data<int, double> md{};
 //   pmap<int, double> m(md);
@@ -58,13 +60,13 @@ class pmap : pmap_trivial_check<K, V>
     pvector<Entry> _vec;
 
     // lower_bound: найти индекс первой записи с ключом >= k.
-    unsigned lower_bound(const K& k) const
+    uintptr_t lower_bound(const K& k) const
     {
-        unsigned lo = 0, hi = _vd.size;
+        uintptr_t lo = 0, hi = _vd.size;
         while( lo < hi )
         {
-            unsigned mid = (lo + hi) / 2;
-            if( _vd.data[mid].key < k )
+            uintptr_t mid = (lo + hi) / 2;
+            if( _vd.data[static_cast<unsigned>(mid)].key < k )
                 lo = mid + 1;
             else
                 hi = mid;
@@ -75,53 +77,61 @@ class pmap : pmap_trivial_check<K, V>
 public:
     explicit pmap(pmap_data<K, V>& data) : _vd(data), _vec(data) {}
 
-    unsigned size()  const { return _vd.size; }
-    bool     empty() const { return _vd.size == 0; }
+    uintptr_t size()  const { return _vd.size; }
+    bool      empty() const { return _vd.size == 0; }
 
     // insert: добавить или заменить ключ k со значением v.
     // Поддерживает отсортированный порядок.
     void insert(const K& k, const V& v)
     {
-        unsigned idx = lower_bound(k);
-        if( idx < _vd.size && !( k < _vd.data[idx].key ) && !( _vd.data[idx].key < k ) )
+        uintptr_t idx = lower_bound(k);
+        if( idx < _vd.size
+            && !( k < _vd.data[static_cast<unsigned>(idx)].key )
+            && !( _vd.data[static_cast<unsigned>(idx)].key < k ) )
         {
             // Ключ уже существует — обновляем значение.
-            _vd.data[idx].value = v;
+            _vd.data[static_cast<unsigned>(idx)].value = v;
             return;
         }
         // Вставляем в позицию idx, сдвигая элементы вправо.
         _vec.push_back(Entry{});   // обеспечиваем ёмкость, добавляем заглушку
-        for( unsigned i = _vd.size - 1; i > idx; i-- )
-            _vd.data[i] = _vd.data[i - 1];
-        _vd.data[idx] = Entry{ k, v };
+        for( uintptr_t i = _vd.size - 1; i > idx; i-- )
+            _vd.data[static_cast<unsigned>(i)] = _vd.data[static_cast<unsigned>(i - 1)];
+        _vd.data[static_cast<unsigned>(idx)] = Entry{ k, v };
     }
 
     // find: вернуть указатель на значение по ключу k или nullptr, если не найдено.
     V* find(const K& k)
     {
-        unsigned idx = lower_bound(k);
-        if( idx < _vd.size && !( k < _vd.data[idx].key ) && !( _vd.data[idx].key < k ) )
-            return &_vd.data[idx].value;
+        uintptr_t idx = lower_bound(k);
+        if( idx < _vd.size
+            && !( k < _vd.data[static_cast<unsigned>(idx)].key )
+            && !( _vd.data[static_cast<unsigned>(idx)].key < k ) )
+            return &_vd.data[static_cast<unsigned>(idx)].value;
         return nullptr;
     }
 
     const V* find(const K& k) const
     {
-        unsigned idx = lower_bound(k);
-        if( idx < _vd.size && !( k < _vd.data[idx].key ) && !( _vd.data[idx].key < k ) )
-            return &_vd.data[idx].value;
+        uintptr_t idx = lower_bound(k);
+        if( idx < _vd.size
+            && !( k < _vd.data[static_cast<unsigned>(idx)].key )
+            && !( _vd.data[static_cast<unsigned>(idx)].key < k ) )
+            return &_vd.data[static_cast<unsigned>(idx)].value;
         return nullptr;
     }
 
     // erase: удалить запись с ключом k. Возвращает true, если найдена и удалена.
     bool erase(const K& k)
     {
-        unsigned idx = lower_bound(k);
-        if( idx >= _vd.size || ( k < _vd.data[idx].key ) || ( _vd.data[idx].key < k ) )
+        uintptr_t idx = lower_bound(k);
+        if( idx >= _vd.size
+            || ( k < _vd.data[static_cast<unsigned>(idx)].key )
+            || ( _vd.data[static_cast<unsigned>(idx)].key < k ) )
             return false;
         // Сдвигаем элементы влево.
-        for( unsigned i = idx; i + 1 < _vd.size; i++ )
-            _vd.data[i] = _vd.data[i + 1];
+        for( uintptr_t i = idx; i + 1 < _vd.size; i++ )
+            _vd.data[static_cast<unsigned>(i)] = _vd.data[static_cast<unsigned>(i + 1)];
         _vd.size--;
         return true;
     }
@@ -129,14 +139,16 @@ public:
     // operator[]: вставить значение по умолчанию, если ключ не найден; вернуть ссылку на значение.
     V& operator[](const K& k)
     {
-        unsigned idx = lower_bound(k);
-        if( idx < _vd.size && !( k < _vd.data[idx].key ) && !( _vd.data[idx].key < k ) )
-            return _vd.data[idx].value;
+        uintptr_t idx = lower_bound(k);
+        if( idx < _vd.size
+            && !( k < _vd.data[static_cast<unsigned>(idx)].key )
+            && !( _vd.data[static_cast<unsigned>(idx)].key < k ) )
+            return _vd.data[static_cast<unsigned>(idx)].value;
         // Вставляем значение по умолчанию.
         V def{};
         insert(k, def);
         idx = lower_bound(k);
-        return _vd.data[idx].value;
+        return _vd.data[static_cast<unsigned>(idx)].value;
     }
 
     // clear: удалить все записи. НЕ освобождает выделенный буфер.
@@ -148,12 +160,12 @@ public:
     // Итерация (по объектам Entry в отсортированном по ключу порядке).
     class iterator
     {
-        VecData* _pd;
-        unsigned _idx;
+        VecData*  _pd;
+        uintptr_t _idx;
     public:
-        iterator(VecData* pd, unsigned idx) : _pd(pd), _idx(idx) {}
-        Entry& operator*()  { return _pd->data[_idx]; }
-        Entry* operator->() { return &_pd->data[_idx]; }
+        iterator(VecData* pd, uintptr_t idx) : _pd(pd), _idx(idx) {}
+        Entry& operator*()  { return _pd->data[static_cast<unsigned>(_idx)]; }
+        Entry* operator->() { return &_pd->data[static_cast<unsigned>(_idx)]; }
         iterator& operator++() { ++_idx; return *this; }
         iterator  operator++(int) { iterator tmp = *this; ++_idx; return tmp; }
         bool operator==(const iterator& o) const { return _idx == o._idx; }
@@ -163,11 +175,11 @@ public:
     class const_iterator
     {
         const VecData* _pd;
-        unsigned       _idx;
+        uintptr_t      _idx;
     public:
-        const_iterator(const VecData* pd, unsigned idx) : _pd(pd), _idx(idx) {}
-        const Entry& operator*()  const { return _pd->data[_idx]; }
-        const Entry* operator->() const { return &_pd->data[_idx]; }
+        const_iterator(const VecData* pd, uintptr_t idx) : _pd(pd), _idx(idx) {}
+        const Entry& operator*()  const { return _pd->data[static_cast<unsigned>(_idx)]; }
+        const Entry* operator->() const { return &_pd->data[static_cast<unsigned>(_idx)]; }
         const_iterator& operator++() { ++_idx; return *this; }
         const_iterator  operator++(int) { const_iterator tmp = *this; ++_idx; return tmp; }
         bool operator==(const const_iterator& o) const { return _idx == o._idx; }
